Rename app
Why:
We would like for the app name to arcade chess and for that name to
appear on the splashboard and in the app store.

This commit:
was generated by using the react-native-rename library.
