+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.realtimechess",
+    package = "com.arcadechess",
 )
 
 android_resource(
     name = "res",
-    package = "com.realtimechess",
+    package = "com.arcadechess",
     res = "src/main/res",
 )
