+++ android/app/src/main/java/com/arcadechess/MainApplication.java
@@ -1,4 +1,4 @@
-package com.realtimechess;
+package com.arcadechess;
 
 import android.app.Application;
 import android.util.Log;
