+++ app.json
@@ -1,4 +1,4 @@
 {
-  "displayName": "RealTimeChess",
-  "name": "RealTimeChess"
+  "displayName": "Arcade Chess",
+  "name": "ArcadeChess"
 }
