+++ ios/ArcadeChess-tvOS/Info.plist
