+++ ios/ArcadeChess.xcodeproj/xcshareddata/xcschemes/ArcadeChess-tvOS.xcscheme
@@ -14,9 +14,9 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "83CBBA2D1A601D0E00E9B192"
+               BlueprintIdentifier = "2D2A28121D9B038B00D4039D"
                BuildableName = "libReact.a"
-               BlueprintName = "React"
+               BlueprintName = "React-tvOS"
                ReferencedContainer = "container:../node_modules/react-native/React/React.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -28,10 +28,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "RealTimeChess.app"
-               BlueprintName = "RealTimeChess"
-               ReferencedContainer = "container:RealTimeChess.xcodeproj">
+               BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+               BuildableName = "ArcadeChess-tvOS.app"
+               BlueprintName = "ArcadeChess-tvOS"
+               ReferencedContainer = "container:realtimechess.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -42,10 +42,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "RealTimeChessTests.xctest"
-               BlueprintName = "RealTimeChessTests"
-               ReferencedContainer = "container:RealTimeChess.xcodeproj">
+               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
+               BuildableName = "ArcadeChess-tvOSTests.xctest"
+               BlueprintName = "ArcadeChess-tvOSTests"
+               ReferencedContainer = "container:realtimechess.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -58,10 +58,10 @@
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RealTimeChess.app"
-            BlueprintName = "RealTimeChess"
-            ReferencedContainer = "container:RealTimeChess.xcodeproj">
+            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+            BuildableName = "ArcadeChess-tvOS.app"
+            BlueprintName = "ArcadeChess-tvOS"
+            ReferencedContainer = "container:realtimechess.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <Testables>
@@ -69,10 +69,10 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "RealTimeChessTests.xctest"
-               BlueprintName = "RealTimeChessTests"
-               ReferencedContainer = "container:RealTimeChess.xcodeproj">
+               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
+               BuildableName = "ArcadeChess-tvOSTests.xctest"
+               BlueprintName = "ArcadeChess-tvOSTests"
+               ReferencedContainer = "container:realtimechess.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -91,10 +91,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RealTimeChess.app"
-            BlueprintName = "RealTimeChess"
-            ReferencedContainer = "container:RealTimeChess.xcodeproj">
+            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+            BuildableName = "ArcadeChess-tvOS.app"
+            BlueprintName = "ArcadeChess-tvOS"
+            ReferencedContainer = "container:realtimechess.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -108,10 +108,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RealTimeChess.app"
-            BlueprintName = "RealTimeChess"
-            ReferencedContainer = "container:RealTimeChess.xcodeproj">
+            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+            BuildableName = "ArcadeChess-tvOS.app"
+            BlueprintName = "ArcadeChess-tvOS"
+            ReferencedContainer = "container:realtimechess.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
