+++ ios/ArcadeChess.xcodeproj/xcshareddata/xcschemes/ArcadeChess.xcscheme
@@ -14,9 +14,9 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D2A28121D9B038B00D4039D"
+               BlueprintIdentifier = "83CBBA2D1A601D0E00E9B192"
                BuildableName = "libReact.a"
-               BlueprintName = "React-tvOS"
+               BlueprintName = "React"
                ReferencedContainer = "container:../node_modules/react-native/React/React.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -28,10 +28,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "RealTimeChess-tvOS.app"
-               BlueprintName = "RealTimeChess-tvOS"
-               ReferencedContainer = "container:realtimechess.xcodeproj">
+               BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+               BuildableName = "ArcadeChess.app"
+               BlueprintName = "ArcadeChess"
+               ReferencedContainer = "container:ArcadeChess.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -42,10 +42,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "RealTimeChess-tvOSTests.xctest"
-               BlueprintName = "RealTimeChess-tvOSTests"
-               ReferencedContainer = "container:realtimechess.xcodeproj">
+               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
+               BuildableName = "ArcadeChessTests.xctest"
+               BlueprintName = "ArcadeChessTests"
+               ReferencedContainer = "container:ArcadeChess.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -58,10 +58,10 @@
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "RealTimeChess-tvOS.app"
-            BlueprintName = "RealTimeChess-tvOS"
-            ReferencedContainer = "container:realtimechess.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "ArcadeChess.app"
+            BlueprintName = "ArcadeChess"
+            ReferencedContainer = "container:ArcadeChess.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <Testables>
@@ -69,10 +69,10 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "RealTimeChess-tvOSTests.xctest"
-               BlueprintName = "RealTimeChess-tvOSTests"
-               ReferencedContainer = "container:realtimechess.xcodeproj">
+               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
+               BuildableName = "ArcadeChessTests.xctest"
+               BlueprintName = "ArcadeChessTests"
+               ReferencedContainer = "container:ArcadeChess.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -91,10 +91,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "RealTimeChess-tvOS.app"
-            BlueprintName = "RealTimeChess-tvOS"
-            ReferencedContainer = "container:realtimechess.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "ArcadeChess.app"
+            BlueprintName = "ArcadeChess"
+            ReferencedContainer = "container:ArcadeChess.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -108,10 +108,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "RealTimeChess-tvOS.app"
-            BlueprintName = "RealTimeChess-tvOS"
-            ReferencedContainer = "container:realtimechess.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "ArcadeChess.app"
+            BlueprintName = "ArcadeChess"
+            ReferencedContainer = "container:ArcadeChess.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
