+++ ios/ArcadeChess/AppDelegate.h
