+++ ios/ArcadeChess/Images.xcassets/AppIcon.appiconset/1024x1.jpg
