+++ ios/ArcadeChess/Images.xcassets/AppIcon.appiconset/20x2.jpg
