+++ ios/ArcadeChess/Images.xcassets/AppIcon.appiconset/20x3.jpg
