+++ ios/ArcadeChess/Images.xcassets/AppIcon.appiconset/29x2.jpg
