+++ ios/ArcadeChess/Images.xcassets/AppIcon.appiconset/29x3.jpg
