+++ ios/ArcadeChess/Images.xcassets/AppIcon.appiconset/40x2.jpg
