+++ ios/ArcadeChess/Images.xcassets/AppIcon.appiconset/40x3.jpg
