+++ ios/ArcadeChess/Images.xcassets/AppIcon.appiconset/60x2.jpg
