+++ ios/ArcadeChess/Images.xcassets/AppIcon.appiconset/60x3.jpg
