+++ ios/ArcadeChess/Images.xcassets/AppIcon.appiconset/Contents.json
