+++ ios/ArcadeChess/Images.xcassets/Contents.json
