+++ ios/ArcadeChess/Images.xcassets/SplashBg.imageset/splash_bgx1.png
