+++ ios/ArcadeChess/Images.xcassets/SplashBg.imageset/splash_bgx2.png
