+++ ios/ArcadeChess/Images.xcassets/SplashBg.imageset/splash_bgx3.png
