+++ ios/ArcadeChess/Info.plist
