+++ ios/ArcadeChessTests/ArcadeChessTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface RealTimeChessTests : XCTestCase
+@interface ArcadeChessTests : XCTestCase
 
 @end
 
-@implementation RealTimeChessTests
+@implementation ArcadeChessTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
