+++ ios/ArcadeChessTests/Info.plist
